Fix a type mismatch introduced by the merge commit e0a1c745ec3ed1ec6c0375a2a624697c29f480a6

+++ sql/log_event.cc
@@ -8673,7 +8673,7 @@ User_var_log_event(const char* buf, uint event_len,
       Old events will not have this extra byte, thence,
       we keep the flags set to UNDEF_F.
     */
-    uint bytes_read= (uint)((val + val_len) - buf_start);
+    size_t bytes_read= ((val + val_len) - buf_start);
     if ((data_written - bytes_read) > 0)
     {
       flags= (uint) *(buf + UV_VAL_IS_NULL + UV_VAL_TYPE_SIZE +
